+++ Tests/AzureFunctions.TestApplication/AzureFunctions.TestApplication.Api/FindSampleDomains.cs
@@ -4,7 +4,6 @@ using System.IO;
 using System.Net;
 using System.Threading;
 using System.Threading.Tasks;
-using AzureFunctions.TestApplication.Application.Common.Pagination;
 using AzureFunctions.TestApplication.Application.Interfaces;
 using AzureFunctions.TestApplication.Application.SampleDomains;
 using AzureFunctions.TestApplication.Domain.Common.Exceptions;
@@ -21,30 +20,26 @@ using Microsoft.OpenApi.Models;
 
 namespace AzureFunctions.TestApplication.Api
 {
-    public class FindSampleDomainsPaged
+    public class FindSampleDomains
     {
         private readonly ISampleDomainsService _appService;
 
-        public FindSampleDomainsPaged(ISampleDomainsService appService)
+        public FindSampleDomains(ISampleDomainsService appService)
         {
             _appService = appService ?? throw new ArgumentNullException(nameof(appService));
         }
 
-        [FunctionName("FindSampleDomainsPaged")]
-        [OpenApiOperation("FindSampleDomainsPaged", tags: new[] { "SampleDomainsPaged" }, Description = "Find sample domains paged")]
-        [OpenApiParameter(name: "pageNo", In = ParameterLocation.Query, Required = true, Type = typeof(int))]
-        [OpenApiParameter(name: "pageSize", In = ParameterLocation.Query, Required = true, Type = typeof(int))]
-        [OpenApiResponseWithBody(statusCode: HttpStatusCode.OK, contentType: "application/json", bodyType: typeof(PagedResult<SampleDomainDto>))]
+        [FunctionName("FindSampleDomains")]
+        [OpenApiOperation("FindSampleDomains", tags: new[] { "SampleDomains" }, Description = "Find sample domains")]
+        [OpenApiResponseWithBody(statusCode: HttpStatusCode.OK, contentType: "application/json", bodyType: typeof(List<SampleDomainDto>))]
         [OpenApiResponseWithBody(statusCode: HttpStatusCode.BadRequest, contentType: "application/json", bodyType: typeof(object))]
         public async Task<IActionResult> Run(
-            [HttpTrigger(AuthorizationLevel.Function, "get", Route = "sample-domains-paged")] HttpRequest req,
+            [HttpTrigger(AuthorizationLevel.Function, "get", Route = "sample-domains")] HttpRequest req,
             CancellationToken cancellationToken)
         {
             try
             {
-                int pageNo = AzureFunctionHelper.GetQueryParam("pageNo", req.Query, (string val, out int parsed) => int.TryParse(val, out parsed));
-                int pageSize = AzureFunctionHelper.GetQueryParam("pageSize", req.Query, (string val, out int parsed) => int.TryParse(val, out parsed));
-                var result = await _appService.FindSampleDomainsPaged(pageNo, pageSize, cancellationToken);
+                var result = await _appService.FindSampleDomains();
                 return new OkObjectResult(result);
             }
             catch (NotFoundException exception)